@@ -1,3 +1,4 @@
+
 package org.springframework.samples.petclinic.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -8,14 +9,13 @@ import java.util.Set;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validator;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 
 /**
- * @author Michael Isvy
- *         Simple test to make sure that Bean Validation is working
- *         (useful when upgrading to a new version of Hibernate Validator/ Bean Validation)
+ * Simple test to make sure that Bean Validation is working
+ * (useful when upgrading to a new version of Hibernate Validator/Bean Validation)
  */
 public class ValidatorTests {
 
@@ -27,8 +27,8 @@ public class ValidatorTests {
 
     @Test
     public void shouldNotValidateWhenFirstNameEmpty() {
-
         LocaleContextHolder.setLocale(Locale.ENGLISH);
+
         Person person = new Person();
         person.setFirstName("");
         person.setLastName("smith");
@@ -39,7 +39,7 @@ public class ValidatorTests {
         assertThat(constraintViolations.size()).isEqualTo(1);
         ConstraintViolation<Person> violation = constraintViolations.iterator().next();
         assertThat(violation.getPropertyPath().toString()).isEqualTo("firstName");
+        // Mensagem padrão atual do @NotEmpty em inglês:
         assertThat(violation.getMessage()).isEqualTo("may not be empty");
     }
-
 }
