@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+
 package org.springframework.samples.petclinic.rest;
 
 import static org.mockito.BDDMockito.given;
@@ -28,195 +29,216 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.samples.petclinic.model.Specialty;
-import org.springframework.samples.petclinic.service.clinicService.ApplicationTestConfig;
 import org.springframework.samples.petclinic.service.ClinicService;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
  * Test class for {@link SpecialtyRestController}
  *
- * @author Vitaliy Fedoriv
+ * Ajustado para JUnit 5 (Jupiter) e compatibilidade com Spring Boot 3.x.
  */
 @SpringBootTest
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes=ApplicationTestConfig.class)
-@WebAppConfiguration
 public class SpecialtyRestControllerTests {
 
     @Autowired
     private SpecialtyRestController specialtyRestController;
 
-	@MockBean
+    @MockBean
     private ClinicService clinicService;
+
+    @Autowired
+    private ObjectMapper objectMapper;
 
     private MockMvc mockMvc;
 
     private List<Specialty> specialties;
 
-    @Before
-    public void initSpecialtys(){
-    	this.mockMvc = MockMvcBuilders.standaloneSetup(specialtyRestController)
-    			.setControllerAdvice(new ExceptionControllerAdvice())
-    			.build();
-    	specialties = new ArrayList<Specialty>();
+    @BeforeEach
+    public void initSpecialtys() {
+        this.mockMvc = MockMvcBuilders.standaloneSetup(specialtyRestController)
+            .setControllerAdvice(new ExceptionControllerAdvice())
+            .build();
 
-    	Specialty specialty = new Specialty();
-    	specialty.setId(1);
-    	specialty.setName("radiology");
-    	specialties.add(specialty);
+        specialties = new ArrayList<>();
 
-    	specialty = new Specialty();
-    	specialty.setId(2);
-    	specialty.setName("surgery");
-    	specialties.add(specialty);
+        Specialty specialty = new Specialty();
+        specialty.setId(1);
+        specialty.setName("radiology");
+        specialties.add(specialty);
 
-    	specialty = new Specialty();
-    	specialty.setId(3);
-    	specialty.setName("dentistry");
-    	specialties.add(specialty);
+        specialty = new Specialty();
+        specialty.setId(2);
+        specialty.setName("surgery");
+        specialties.add(specialty);
 
+        specialty = new Specialty();
+        specialty.setId(3);
+        specialty.setName("dentistry");
+        specialties.add(specialty);
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetSpecialtySuccess() throws Exception {
-    	given(this.clinicService.findSpecialtyById(1)).willReturn(specialties.get(0));
+        given(this.clinicService.findSpecialtyById(1)).willReturn(specialties.get(0));
+
         this.mockMvc.perform(get("/api/specialties/1")
-        	.accept(MediaType.APPLICATION_JSON_VALUE))
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.id").value(1))
             .andExpect(jsonPath("$.name").value("radiology"));
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetSpecialtyNotFound() throws Exception {
-    	given(this.clinicService.findSpecialtyById(-1)).willReturn(null);
+        given(this.clinicService.findSpecialtyById(-1)).willReturn(null);
+
         this.mockMvc.perform(get("/api/specialties/-1")
-        	.accept(MediaType.APPLICATION_JSON))
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetAllSpecialtysSuccess() throws Exception {
-    	specialties.remove(0);
-    	given(this.clinicService.findAllSpecialties()).willReturn(specialties);
-        this.mockMvc.perform(get("/api/specialties/")
-        	.accept(MediaType.APPLICATION_JSON))
+        // Remove o primeiro (id 1), mantendo id 2 e 3
+        specialties.remove(0);
+        given(this.clinicService.findAllSpecialties()).willReturn(specialties);
+
+        this.mockMvc.perform(get("/api/specialties")
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
-        	.andExpect(jsonPath("$.[0].id").value(2))
-        	.andExpect(jsonPath("$.[0].name").value("surgery"))
-        	.andExpect(jsonPath("$.[1].id").value(3))
-        	.andExpect(jsonPath("$.[1].name").value("dentistry"));
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.[0].id").value(2))
+            .andExpect(jsonPath("$.[0].name").value("surgery"))
+            .andExpect(jsonPath("$.[1].id").value(3))
+            .andExpect(jsonPath("$.[1].name").value("dentistry"));
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetAllSpecialtysNotFound() throws Exception {
-    	specialties.clear();
-    	given(this.clinicService.findAllSpecialties()).willReturn(specialties);
-        this.mockMvc.perform(get("/api/specialties/")
-        	.accept(MediaType.APPLICATION_JSON))
+        specialties.clear();
+        given(this.clinicService.findAllSpecialties()).willReturn(specialties);
+
+        this.mockMvc.perform(get("/api/specialties")
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testCreateSpecialtySuccess() throws Exception {
-    	Specialty newSpecialty = specialties.get(0);
-    	newSpecialty.setId(999);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newSpecialtyAsJSON = mapper.writeValueAsString(newSpecialty);
-    	this.mockMvc.perform(post("/api/specialties/")
-    		.content(newSpecialtyAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-    		.andExpect(status().isCreated());
+        Specialty newSpecialty = specialties.get(0);
+        newSpecialty.setId(999);
+
+        String newSpecialtyAsJSON = objectMapper.writeValueAsString(newSpecialty);
+
+        this.mockMvc.perform(post("/api/specialties")
+                .content(newSpecialtyAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isCreated());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testCreateSpecialtyError() throws Exception {
-    	Specialty newSpecialty = specialties.get(0);
-    	newSpecialty.setId(null);
-    	newSpecialty.setName(null);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newSpecialtyAsJSON = mapper.writeValueAsString(newSpecialty);
-    	this.mockMvc.perform(post("/api/specialties/")
-        		.content(newSpecialtyAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        		.andExpect(status().isBadRequest());
-     }
+        Specialty newSpecialty = specialties.get(0);
+        newSpecialty.setId(null);
+        newSpecialty.setName(null);
+
+        String newSpecialtyAsJSON = objectMapper.writeValueAsString(newSpecialty);
+
+        this.mockMvc.perform(post("/api/specialties")
+                .content(newSpecialtyAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest());
+    }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testUpdateSpecialtySuccess() throws Exception {
-    	given(this.clinicService.findSpecialtyById(2)).willReturn(specialties.get(1));
-    	Specialty newSpecialty = specialties.get(1);
-    	newSpecialty.setName("surgery I");
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newSpecialtyAsJSON = mapper.writeValueAsString(newSpecialty);
-    	this.mockMvc.perform(put("/api/specialties/2")
-    		.content(newSpecialtyAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(content().contentType("application/json;charset=UTF-8"))
-        	.andExpect(status().isNoContent());
+        given(this.clinicService.findSpecialtyById(2)).willReturn(specialties.get(1));
 
-    	this.mockMvc.perform(get("/api/specialties/2")
-           	.accept(MediaType.APPLICATION_JSON).contentType(MediaType.APPLICATION_JSON_VALUE))
+        Specialty newSpecialty = specialties.get(1);
+        newSpecialty.setName("surgery I");
+
+        String newSpecialtyAsJSON = objectMapper.writeValueAsString(newSpecialty);
+
+        this.mockMvc.perform(put("/api/specialties/2")
+                .content(newSpecialtyAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNoContent());
+
+        this.mockMvc.perform(get("/api/specialties/2")
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.id").value(2))
             .andExpect(jsonPath("$.name").value("surgery I"));
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testUpdateSpecialtyError() throws Exception {
-    	Specialty newSpecialty = specialties.get(0);
-    	newSpecialty.setName("");
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newSpecialtyAsJSON = mapper.writeValueAsString(newSpecialty);
-    	this.mockMvc.perform(put("/api/specialties/1")
-    		.content(newSpecialtyAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isBadRequest());
-     }
+        Specialty newSpecialty = specialties.get(0);
+        newSpecialty.setName("");
 
-    @Test
-    @WithMockUser(roles="VET_ADMIN")
-    public void testDeleteSpecialtySuccess() throws Exception {
-    	Specialty newSpecialty = specialties.get(0);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newSpecialtyAsJSON = mapper.writeValueAsString(newSpecialty);
-    	given(this.clinicService.findSpecialtyById(1)).willReturn(specialties.get(0));
-    	this.mockMvc.perform(delete("/api/specialties/1")
-    		.content(newSpecialtyAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isNoContent());
+        String newSpecialtyAsJSON = objectMapper.writeValueAsString(newSpecialty);
+
+        this.mockMvc.perform(put("/api/specialties/1")
+                .content(newSpecialtyAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
+    public void testDeleteSpecialtySuccess() throws Exception {
+        Specialty newSpecialty = specialties.get(0);
+        String newSpecialtyAsJSON = objectMapper.writeValueAsString(newSpecialty);
+
+        given(this.clinicService.findSpecialtyById(1)).willReturn(specialties.get(0));
+
+        this.mockMvc.perform(delete("/api/specialties/1")
+                .content(newSpecialtyAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNoContent());
+    }
+
+    @Test
+    @WithMockUser(roles = "VET_ADMIN")
     public void testDeleteSpecialtyError() throws Exception {
-    	Specialty newSpecialty = specialties.get(0);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newSpecialtyAsJSON = mapper.writeValueAsString(newSpecialty);
-    	given(this.clinicService.findSpecialtyById(-1)).willReturn(null);
-    	this.mockMvc.perform(delete("/api/specialties/-1")
-    		.content(newSpecialtyAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isNotFound());
+        Specialty newSpecialty = specialties.get(0);
+        String newSpecialtyAsJSON = objectMapper.writeValueAsString(newSpecialty);
+
+        given(this.clinicService.findSpecialtyById(-1)).willReturn(null);
+
+        this.mockMvc.perform(delete("/api/specialties/-1")
+                .content(newSpecialtyAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNotFound());
     }
 }
