@@ -14,116 +14,114 @@
  * limitations under the License.
  */
 
+
 package org.springframework.samples.petclinic.rest;
 
 import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.samples.petclinic.model.Vet;
-import org.springframework.samples.petclinic.service.clinicService.ApplicationTestConfig;
 import org.springframework.samples.petclinic.service.ClinicService;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
  * Test class for {@link VetRestController}
  *
+ * Ajustado para JUnit 5 (Jupiter) e compatibilidade com Spring Boot 3.x.
+ *
  * @author Vitaliy Fedoriv
  */
+
 @SpringBootTest
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes=ApplicationTestConfig.class)
-@WebAppConfiguration
 public class VetRestControllerTests {
 
     @Autowired
     private VetRestController vetRestController;
 
-	@MockBean
+    @MockBean
     private ClinicService clinicService;
+
+    @Autowired
+    private ObjectMapper objectMapper;
 
     private MockMvc mockMvc;
 
     private List<Vet> vets;
 
-    @Before
-    public void initVets(){
-    	this.mockMvc = MockMvcBuilders.standaloneSetup(vetRestController)
-    			.setControllerAdvice(new ExceptionControllerAdvice())
-    			.build();
-    	vets = new ArrayList<Vet>();
+    @BeforeEach
+    public void initVets() {
+        this.mockMvc = MockMvcBuilders.standaloneSetup(vetRestController)
+            .setControllerAdvice(new ExceptionControllerAdvice())
+            .build();
 
+        vets = new ArrayList<>();
 
-    	Vet vet = new Vet();
-    	vet.setId(1);
-    	vet.setFirstName("James");
-    	vet.setLastName("Carter");
-    	vets.add(vet);
+        Vet vet = new Vet();
+        vet.setId(1);
+        vet.setFirstName("James");
+        vet.setLastName("Carter");
+        vets.add(vet);
 
-    	vet = new Vet();
-    	vet.setId(2);
-    	vet.setFirstName("Helen");
-    	vet.setLastName("Leary");
-    	vets.add(vet);
+        vet = new Vet();
+        vet.setId(2);
+        vet.setFirstName("Helen");
+        vet.setLastName("Leary");
+        vets.add(vet);
 
-    	vet = new Vet();
-    	vet.setId(3);
-    	vet.setFirstName("Linda");
-    	vet.setLastName("Douglas");
-    	vets.add(vet);
+        vet = new Vet();
+        vet.setId(3);
+        vet.setFirstName("Linda");
+        vet.setLastName("Douglas");
+        vets.add(vet);
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetVetSuccess() throws Exception {
-    	given(this.clinicService.findVetById(1)).willReturn(vets.get(0));
+        given(this.clinicService.findVetById(1)).willReturn(vets.get(0));
+
         this.mockMvc.perform(get("/api/vets/1")
-        	.accept(MediaType.APPLICATION_JSON_VALUE))
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.id").value(1))
             .andExpect(jsonPath("$.firstName").value("James"));
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetVetNotFound() throws Exception {
-    	given(this.clinicService.findVetById(-1)).willReturn(null);
+        given(this.clinicService.findVetById(-1)).willReturn(null);
+
         this.mockMvc.perform(get("/api/vets/-1")
-        	.accept(MediaType.APPLICATION_JSON))
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetAllVetsSuccess() throws Exception {
-    	given(this.clinicService.findAllVets()).willReturn(vets);
-        this.mockMvc.perform(get("/api/vets/")
-        	.accept(MediaType.APPLICATION_JSON))
+        given(this.clinicService.findAllVets()).willReturn(vets);
+
+        this.mockMvc.perform(get("/api/vets")
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.[0].id").value(1))
             .andExpect(jsonPath("$.[0].firstName").value("James"))
             .andExpect(jsonPath("$.[1].id").value(2))
@@ -131,97 +129,115 @@ public class VetRestControllerTests {
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testGetAllVetsNotFound() throws Exception {
-    	vets.clear();
-    	given(this.clinicService.findAllVets()).willReturn(vets);
-        this.mockMvc.perform(get("/api/vets/")
-        	.accept(MediaType.APPLICATION_JSON))
+        vets.clear();
+        given(this.clinicService.findAllVets()).willReturn(vets);
+
+        this.mockMvc.perform(get("/api/vets")
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testCreateVetSuccess() throws Exception {
-    	Vet newVet = vets.get(0);
-    	newVet.setId(999);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVetAsJSON = mapper.writeValueAsString(newVet);
-    	this.mockMvc.perform(post("/api/vets/")
-    		.content(newVetAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-    		.andExpect(status().isCreated());
+        Vet newVet = vets.get(0);
+        newVet.setId(999);
+
+        String newVetAsJSON = objectMapper.writeValueAsString(newVet);
+
+        this.mockMvc.perform(post("/api/vets")
+                .content(newVetAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isCreated());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testCreateVetError() throws Exception {
-    	Vet newVet = vets.get(0);
-    	newVet.setId(null);
-    	newVet.setFirstName(null);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVetAsJSON = mapper.writeValueAsString(newVet);
-    	this.mockMvc.perform(post("/api/vets/")
-        		.content(newVetAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        		.andExpect(status().isBadRequest());
-     }
+        Vet newVet = vets.get(0);
+        newVet.setId(null);
+        newVet.setFirstName(null);
+
+        String newVetAsJSON = objectMapper.writeValueAsString(newVet);
+
+        this.mockMvc.perform(post("/api/vets")
+                .content(newVetAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest());
+    }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testUpdateVetSuccess() throws Exception {
-    	given(this.clinicService.findVetById(1)).willReturn(vets.get(0));
-    	Vet newVet = vets.get(0);
-    	newVet.setFirstName("James");
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVetAsJSON = mapper.writeValueAsString(newVet);
-    	this.mockMvc.perform(put("/api/vets/1")
-    		.content(newVetAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(content().contentType("application/json;charset=UTF-8"))
-        	.andExpect(status().isNoContent());
+        given(this.clinicService.findVetById(1)).willReturn(vets.get(0));
 
-    	this.mockMvc.perform(get("/api/vets/1")
-           	.accept(MediaType.APPLICATION_JSON).contentType(MediaType.APPLICATION_JSON_VALUE))
+        Vet newVet = vets.get(0);
+        newVet.setFirstName("James");
+
+        String newVetAsJSON = objectMapper.writeValueAsString(newVet);
+
+        this.mockMvc.perform(put("/api/vets/1")
+                .content(newVetAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+            .andExpect(status().isOk());
+
+        this.mockMvc.perform(get("/api/vets/1")
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.id").value(1))
             .andExpect(jsonPath("$.firstName").value("James"));
-
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testUpdateVetError() throws Exception {
-    	Vet newVet = vets.get(0);
-    	newVet.setFirstName("");
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVetAsJSON = mapper.writeValueAsString(newVet);
-    	this.mockMvc.perform(put("/api/vets/1")
-    		.content(newVetAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isBadRequest());
-     }
+        Vet newVet = vets.get(0);
+        newVet.setFirstName("");
+
+        String newVetAsJSON = objectMapper.writeValueAsString(newVet);
+
+        this.mockMvc.perform(put("/api/vets/1")
+                .content(newVetAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest());
+    }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testDeleteVetSuccess() throws Exception {
-    	Vet newVet = vets.get(0);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVetAsJSON = mapper.writeValueAsString(newVet);
-    	given(this.clinicService.findVetById(1)).willReturn(vets.get(0));
-    	this.mockMvc.perform(delete("/api/vets/1")
-    		.content(newVetAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isNoContent());
+        Vet newVet = vets.get(0);
+        String newVetAsJSON = objectMapper.writeValueAsString(newVet);
+
+        given(this.clinicService.findVetById(1)).willReturn(vets.get(0));
+
+        this.mockMvc.perform(delete("/api/vets/1")
+                .content(newVetAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNoContent());
     }
 
     @Test
-    @WithMockUser(roles="VET_ADMIN")
+    @WithMockUser(roles = "VET_ADMIN")
     public void testDeleteVetError() throws Exception {
-    	Vet newVet = vets.get(0);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVetAsJSON = mapper.writeValueAsString(newVet);
-    	given(this.clinicService.findVetById(-1)).willReturn(null);
-    	this.mockMvc.perform(delete("/api/vets/-1")
-    		.content(newVetAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isNotFound());
+        Vet newVet = vets.get(0);
+        String newVetAsJSON = objectMapper.writeValueAsString(newVet);
+
+        given(this.clinicService.findVetById(-1)).willReturn(null);
+
+        this.mockMvc.perform(delete("/api/vets/-1")
+                .content(newVetAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNotFound());
     }
-
 }
-
