@@ -25,14 +25,15 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -41,26 +42,12 @@ import org.springframework.samples.petclinic.model.Owner;
 import org.springframework.samples.petclinic.model.Pet;
 import org.springframework.samples.petclinic.model.PetType;
 import org.springframework.samples.petclinic.model.Visit;
-import org.springframework.samples.petclinic.service.clinicService.ApplicationTestConfig;
 import org.springframework.samples.petclinic.service.ClinicService;
 import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-/**
- * Test class for {@link VisitRestController}
- *
- * @author Vitaliy Fedoriv
- */
 @SpringBootTest
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes=ApplicationTestConfig.class)
-@WebAppConfiguration
 public class VisitRestControllerTests {
 
     @Autowired
@@ -69,183 +56,191 @@ public class VisitRestControllerTests {
     @MockBean
     private ClinicService clinicService;
 
+    @Autowired
+    private ObjectMapper objectMapper;
+
     private MockMvc mockMvc;
 
     private List<Visit> visits;
 
-    @Before
-    public void initVisits(){
-    	this.mockMvc = MockMvcBuilders.standaloneSetup(visitRestController)
-    			.setControllerAdvice(new ExceptionControllerAdvice())
-    			.build();
+    @BeforeEach
+    public void initVisits() {
+        this.mockMvc = MockMvcBuilders.standaloneSetup(visitRestController)
+            .setControllerAdvice(new ExceptionControllerAdvice())
+            .build();
 
-    	visits = new ArrayList<Visit>();
+        visits = new ArrayList<>();
 
-    	Owner owner = new Owner();
-    	owner.setId(1);
-    	owner.setFirstName("Eduardo");
-    	owner.setLastName("Rodriquez");
-    	owner.setAddress("2693 Commerce St.");
-    	owner.setCity("McFarland");
+        Owner owner = new Owner();
+        owner.setId(1);
+        owner.setFirstName("Eduardo");
+        owner.setLastName("Rodriquez");
+        owner.setAddress("2693 Commerce St.");
+        owner.setCity("McFarland");
         owner.setState("Wisconsin");
         owner.setZipCode("53558");
-    	owner.setTelephone("6085558763");
+        owner.setTelephone("6085558763");
 
-    	PetType petType = new PetType();
-    	petType.setId(2);
-    	petType.setName("dog");
+        PetType petType = new PetType();
+        petType.setId(2);
+        petType.setName("dog");
 
-    	Pet pet = new Pet();
-    	pet.setId(8);
-    	pet.setName("Rosy");
-    	pet.setBirthDate(new Date());
-    	pet.setOwner(owner);
-    	pet.setType(petType);
+        Pet pet = new Pet();
+        pet.setId(8);
+        pet.setName("Rosy");
+        pet.setBirthDate(new Date());
+        pet.setOwner(owner);
+        pet.setType(petType);
 
+        Visit visit = new Visit();
+        visit.setId(2);
+        visit.setPet(pet);
+        visit.setDate(new Date());
+        visit.setDescription("rabies shot");
+        visits.add(visit);
 
-    	Visit visit = new Visit();
-    	visit.setId(2);
-    	visit.setPet(pet);
-    	visit.setDate(new Date());
-    	visit.setDescription("rabies shot");
-    	visits.add(visit);
-
-    	visit = new Visit();
-    	visit.setId(3);
-    	visit.setPet(pet);
-    	visit.setDate(new Date());
-    	visit.setDescription("neutered");
-    	visits.add(visit);
-
-
+        visit = new Visit();
+        visit.setId(3);
+        visit.setPet(pet);
+        visit.setDate(new Date());
+        visit.setDescription("neutered");
+        visits.add(visit);
     }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testGetVisitSuccess() throws Exception {
-    	given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
+        given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
+
         this.mockMvc.perform(get("/api/visits/2")
-        	.accept(MediaType.APPLICATION_JSON_VALUE))
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.id").value(2))
             .andExpect(jsonPath("$.description").value("rabies shot"));
     }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testGetVisitNotFound() throws Exception {
-    	given(this.clinicService.findVisitById(-1)).willReturn(null);
+        given(this.clinicService.findVisitById(-1)).willReturn(null);
+
         this.mockMvc.perform(get("/api/visits/-1")
-        	.accept(MediaType.APPLICATION_JSON))
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testGetAllVisitsSuccess() throws Exception {
-    	given(this.clinicService.findAllVisits()).willReturn(visits);
-        this.mockMvc.perform(get("/api/visits/")
-        	.accept(MediaType.APPLICATION_JSON))
+        given(this.clinicService.findAllVisits()).willReturn(visits);
+
+        this.mockMvc.perform(get("/api/visits")
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
-        	.andExpect(jsonPath("$.[0].id").value(2))
-        	.andExpect(jsonPath("$.[0].description").value("rabies shot"))
-        	.andExpect(jsonPath("$.[1].id").value(3))
-        	.andExpect(jsonPath("$.[1].description").value("neutered"));
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.[0].id").value(2))
+            .andExpect(jsonPath("$.[0].description").value("rabies shot"))
+            .andExpect(jsonPath("$.[1].id").value(3))
+            .andExpect(jsonPath("$.[1].description").value("neutered"));
     }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testGetAllVisitsNotFound() throws Exception {
-    	visits.clear();
-    	given(this.clinicService.findAllVisits()).willReturn(visits);
-        this.mockMvc.perform(get("/api/visits/")
-        	.accept(MediaType.APPLICATION_JSON))
+        visits.clear();
+        given(this.clinicService.findAllVisits()).willReturn(visits);
+
+        this.mockMvc.perform(get("/api/visits")
+                .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isNotFound());
     }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testCreateVisitSuccess() throws Exception {
-    	Visit newVisit = visits.get(0);
-    	newVisit.setId(999);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVisitAsJSON = mapper.writeValueAsString(newVisit);
-    	System.out.println("newVisitAsJSON " + newVisitAsJSON);
-    	this.mockMvc.perform(post("/api/visits/")
-    		.content(newVisitAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-    		.andExpect(status().isCreated());
+        Visit newVisit = visits.get(0);
+        newVisit.setId(999);
+
+        String newVisitAsJSON = objectMapper.writeValueAsString(newVisit);
+
+        this.mockMvc.perform(post("/api/visits")
+                .content(newVisitAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isCreated());
     }
 
-    @Test(expected = IOException.class)
-    @WithMockUser(roles="OWNER_ADMIN")
+    @Test
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testCreateVisitError() throws Exception {
-    	Visit newVisit = visits.get(0);
-    	newVisit.setId(null);
-    	newVisit.setPet(null);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVisitAsJSON = mapper.writeValueAsString(newVisit);
-    	this.mockMvc.perform(post("/api/visits/")
-        		.content(newVisitAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        		.andExpect(status().isBadRequest());
-     }
+        // Envie JSON inválido SEM o campo "pet" para provocar BAD_REQUEST no controller
+        String newVisitAsJSON = "{ \"id\": 999, \"description\": \"invalid without pet\" }";
+    
+        this.mockMvc.perform(post("/api/visits")
+                .content(newVisitAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest());
+    }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testUpdateVisitSuccess() throws Exception {
-    	given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
-    	Visit newVisit = visits.get(0);
-    	newVisit.setDescription("rabies shot test");
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVisitAsJSON = mapper.writeValueAsString(newVisit);
-    	this.mockMvc.perform(put("/api/visits/2")
-    		.content(newVisitAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(content().contentType("application/json;charset=UTF-8"))
-        	.andExpect(status().isNoContent());
+        given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
 
-    	this.mockMvc.perform(get("/api/visits/2")
-           	.accept(MediaType.APPLICATION_JSON).contentType(MediaType.APPLICATION_JSON_VALUE))
+        Visit newVisit = visits.get(0);
+        newVisit.setDescription("rabies shot test");
+
+        String newVisitAsJSON = objectMapper.writeValueAsString(newVisit);
+
+        this.mockMvc.perform(put("/api/visits/2")
+                .content(newVisitAsJSON)
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isNoContent()); // 204 sem body
+
+        // Depois do update 204, validamos com GET:
+        this.mockMvc.perform(get("/api/visits/2")
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json;charset=UTF-8"))
+            .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
             .andExpect(jsonPath("$.id").value(2))
             .andExpect(jsonPath("$.description").value("rabies shot test"));
     }
 
-    @Test(expected = IOException.class)
-    @WithMockUser(roles="OWNER_ADMIN")
+    @Test
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testUpdateVisitError() throws Exception {
-    	Visit newVisit = visits.get(0);
-    	newVisit.setPet(null);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVisitAsJSON = mapper.writeValueAsString(newVisit);
-    	this.mockMvc.perform(put("/api/visits/2")
-    		.content(newVisitAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isBadRequest());
-     }
+        // O controller busca a visita antes de validar
+        given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
+    
+        // JSON inválido: sem "pet"
+        String invalidVisitAsJSON = "{ \"id\": 2, \"description\": \"still invalid without pet\" }";
+    
+        this.mockMvc.perform(put("/api/visits/2")
+                .accept(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest());
+    }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testDeleteVisitSuccess() throws Exception {
-    	Visit newVisit = visits.get(0);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVisitAsJSON = mapper.writeValueAsString(newVisit);
-    	given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
-    	this.mockMvc.perform(delete("/api/visits/2")
-    		.content(newVisitAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isNoContent());
+        given(this.clinicService.findVisitById(2)).willReturn(visits.get(0));
+
+        this.mockMvc.perform(delete("/api/visits/2"))
+            .andExpect(status().isNoContent());
     }
 
     @Test
-    @WithMockUser(roles="OWNER_ADMIN")
+    @WithMockUser(roles = "OWNER_ADMIN")
     public void testDeleteVisitError() throws Exception {
-    	Visit newVisit = visits.get(0);
-    	ObjectMapper mapper = new ObjectMapper();
-    	String newVisitAsJSON = mapper.writeValueAsString(newVisit);
-    	given(this.clinicService.findVisitById(-1)).willReturn(null);
-    	this.mockMvc.perform(delete("/api/visits/-1")
-    		.content(newVisitAsJSON).accept(MediaType.APPLICATION_JSON_VALUE).contentType(MediaType.APPLICATION_JSON_VALUE))
-        	.andExpect(status().isNotFound());
-    }
+        given(this.clinicService.findVisitById(-1)).willReturn(null);
 
+        this.mockMvc.perform(delete("/api/visits/-1"))
+            .andExpect(status().isNotFound());
+    }
 }
+
