@@ -1,9 +1,25 @@
+/*
+ * Copyright 2002-2013 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.springframework.samples.petclinic.service.clinicService;
 
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 /**
  * <p> Integration test using the 'Spring Data' profile.
@@ -13,8 +29,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
  */
 
 @SpringBootTest
-@RunWith(SpringJUnit4ClassRunner.class)
-@ActiveProfiles("spring-data-jpa, hsqldb")
+@ExtendWith(SpringExtension.class)
+@ActiveProfiles({"spring-data-jpa", "hsqldb"})
 public class ClinicServiceSpringDataJpaTests extends AbstractClinicServiceTests {
-
 }
